web/app/templates/debug/debugger/requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0

# Development tools
ipython==8.19.0
ipdb==0.13.13
python-dotenv==1.0.0

# Security scanning
bandit==1.7.6
safety==2.3.5
